rts2_status_t and associated changes
git-svn-id: http://svn.code.sf.net/p/rts-2/code/trunk/rts-2@11304 65ab8f4e-f147-0410-b3a9-f14133ecfe55

+++ src/xmlrpc/xmlapi.cpp
@@ -303,7 +303,7 @@ void DeviceState::sessionExecute (XmlRpcValue& params, XmlRpcValue& result)
 		throw XmlRpcException ("Cannot find specified device");
 	}
 	result[0] = conn->getStateString ();
-	result[1] = conn->getRealState ();
+	result[1] = (int) conn->getRealState ();
 }
 
 void ListValues::sessionExecute (XmlRpcValue& params, XmlRpcValue& result)
